Imported a skeleton portion of map
I am going to approach this map differently; instead of being hard-fixed on what is placed on paper (graphing and grid papers), I am going to allow a bit more of dynamicness.  I am going to just graph the layout - but allow any further changes necessary.  If a part of the map needs to be reconstructed - then do it.