Base updated; seeking outside help - I am stuck here
I have populated this map with a real world scenario where there is a lot of activity going on within the map.  With the map populated in such a way that it is now, how can I easily link the sectors together in a portal fasion in the best case possible?