Added two global vars and CreateNewScriptFile()
Added two variables: SCRIPTFILE - which holds the absolute path of the location in which this script is running from, and SCRIPTFILENAME - which holds the new script file name.

Added a new function: CreateNewScriptFile().  This will generate a new empty file on the secondary storage device within the same directory of the make.ps1 script.