Use $PSScriptRoot instead to get script absolute path
Instead of using the: Split-Path $script:MyInvocation.MyCommand.Path statement in order to get the path of the script, we will use $PSScriptRoot.  This is much cleaner and far easier to understand.  However, PowerShell version 3+ is required.  Though, not much of an issue here as we will focus on version 5.