Further improved CreateNewScriptFile()
Added a new global variable 'PROJECTNAME' which holds the project name.  This will be presented to the user and the compiler script.

CreateNewScriptFile() will now add a header with basic information to the compiler script file.  The header will merely hold the time and date for when the file was generated, but also contains what project the compiler is designed for.