Created the main() function
This function is the main program entry point, thus - this is the primary function for the program.  This function will automatically delete the old script, create the new script, and generate the script.  This function is NOT complete, but it works as of right now.  Obviously, there needs to be further imporvements.