Greatly revised MakeCompiler()
MakeCompiler() has been renamed to MakeCompilerDriver(), MakeCompiler() is now a new function that handles the main task of appending the script file.

MakeCompilerDriver() has been created to simplify the operation in such a way to greatly reduce redundancy in the code, which was clearly apparent - and even more so when DEBUGMODE has been included.