Added a new function WaitUserInput()
This function will prevent the terminal instance from automatically closing after the program reaches the EXIT statement.  This instruction is only useful if the user only openned the terminal instance with the script file only, otherwise - this function is rather unnecessary as the terminal will remain active even after the EXIT statement.