Resolved an issue with relative paths
When the user creates a new instance of PowerShell to run a script directly, the Current Working Directory is in "C:\Windows\System32" but the program will use "C:\" because of the relative paths.  To resolve this, we will have to use Absolute Paths that is relative to the script path.