Created Project Infomration; holds Proj. Information
This script will hold the project's basic information, such as information as: project name, code name, and websites relating to the project and its development.

The variables are to not be tampered with at Run-Time, thus they are supposed to be 'Read Only'.  I will implement get() functions to retrieve the values of the variables properly.  I strongly urge no one to use the variables directly, excluding the member functions within that scope.