Second thought: 'Debug' --> 'Development'
The reason for this change is merely based on that 'Debug' means there is added instructions for the simplication of 'debugging', while 'Development' means that the build could be a candidate for a release.  Though, a Dev. build could contain special instructions for debugging, but the primary purpose is candidates for release.