Added a new implied ExecuteGit() function
The newly added function, ExecuteGit(), will do simply that - execute the git executable, but will require a parameter to determine the operation.
I am still unsure how to have functions be treated as 'private', I have added two underscores at the stem of the function name to imply it is private.