Added LogPath var and revised log paths.
Added a new variable '$__logPath'.  This variable will hold the path for logs based on the initial execution from git.exe.
Revised the '$__reportPath'  to be better organized.
Added a new argument to __ExecuteGit() function, '$logPath'.  This argument will hold where to place the output.  This argument is not yet fully supported.