Big Update: Updated Path Vars & Support Dir. Check
It's been a long while since I've made a commit it seems, thus - this is a big update containing two changes.

First is a small revision on the directory variables.  Added a new variable '$__rootLogPath' which holds the stem of the directory (for example: <PROG>\<Logs>\git).  With this addition, the already existing path variables have been modified to support '$__rootLogPath' instead of using the '$Global:' various times in order to capture the log directory.

Implemented Checking Required Directories and Creating Required Directories.  This implementation will assure that the directories are setup and ready for use.