[WIP] Heavily Revised ExecuteCommand()
Heavily revised the ExecuteCommand() function by using the .NET API to execute a specific command.  WIth this, we can also pass the STDOUT results to a reference parameter variable or throw it in a text file (works in test cases, but not formally ready).

This is very much a WIP; requires code clean up, resolve WriteToFile() function, and proper validations.