Proper support for Report files.
ATTN: Report Filename is no longer automatically generated!  The reason for this change is to assure that with multiple calls to ExecuteCommand() all data is not split in abundance of files, instead - you MUST include a filename, thus preventing that issue.

It is now possible to throw the STDOUT results to a report.