Customizable git.exe paths supported
Added a new variable "$__executablePath" which houses the path of reaching the git.exe binary.  This variable can be useful when the host's environment variable %PATH% does not contain a specific path to git, thus - it is now possible to tell the program where to look to find the executable.  For example, GitHub Desktop throws git.exe "%LocalAppData%GitHubDesktop\app-1.6.0\resources\app\git\mingw64\bin\git.exe" but varies due to different versions.