Check if user allowed update source
Check to make sure that the user has allowed the source to be updated.

NOTE:  I really hate the implementation; I prefer to do:
if (condition0) && (condition1), but where in 'condition0' would fail and jump out of entire conditional statement (like in C), but in PowerShell - the entire conditional statement is executed.  Meaning, if Condition0 fails - Condition1 is still executed.