Resolved issue where the pointers where not storing data.
Resolved an issue where the data stored in the pointers were not properly transfering out-of-the-scopes.  To resolve this, the pointer must point to a memory allocation that is created and bounded by the current scope - in each function.  Thus, the data can not point to a pointer in a pointer, instead it must be a pointer to a valid memory address that is in within that current scope -- memory addresses that were allocated elsewhere are not valid, regardless if used in the arguments.

I really need to stop thinking C....