Added new function: FetchCommitHistory()
Added a new function named FetchCommitHistory().  This function will do essentially as the name implies, retrieves the commit history (aka Change Log).  The commit history will be placed into a text file (as a Report; though for full flexibility as placement and filename) and will be placed in the requested output path.