Fixed a Deadlock issue when executing an extCMD.
Fixed an issue with how the output was previously managed.  Previously the output would remain in the buffer until finally released AFTER the program's execution, but once the buffer is full - a deadlock will occur as nothing more can be pushed into the buffer.  To resolve this, we will immediately cache the results as it is provided by the external command.  With this, we will clear the buffer and also store the output in the cache variables -- this helps to keep the buffer clear and yet still keep the output intact.