Added a new function SwitchLocalBranch()
Added a new function named SwitchLocalBranch().  This function will allow the user to switch from one branch to another, but only if the branch exists in the project's local repository.  This functionality will effect ALL sources, not just the program itself -- if the branch is changed, it will effect all external resources and not just this program alone.  The change is NOT temporary, the user must change the branch again - if so desired to revert back.