Added a new function: FetchAllContributors()
Added a new function named 'FetchAllContributors()'.  This function will allow the ability to retrieve all contributors that have been involved within the project.  The information provided are the following:
 - User's name
 - User's e-mail address
 - How many commits submitted
If the user does NOT have a valid email address nor has made any commits to the project, then that person will not be listed.