Added a new function 'FetchAllBranchesActivity()'
Created a new function named 'FetchAllBranchesActivity()' which is adjacent to 'FetchAllBranches()', but attaches the last-known date of the last commit.  This, in addition, uses the remote repository.  To assure we fetch ALL branches, we will use the centralized repository.  If we were to use the local repository, we may have missing branches that have yet to be downloaded or active in that repository.

CreateNewReport() now uses FetchAllBranchesActivity() instead of FetchAllBranches.