Change PDF page orientation to Landscape
Changed the PDF page orientation to Landscape.  This change is kind-of necessary if the project is rather large and uses a lot of branches.  I can not guarantee if the project is too large that one page (horizontally) will be two or more pages (horizontally) due to abundancy of branches being active all at once.