Revised logging in 7Zip Class
Revised how logging directories are to be managed within this class.  Previously the logging functionality was not yet planned out - or was an oversight during the planning stages.  The changes here are to further ingrain proper logging and report mangement.

Added two new variables:  '$__rootLogPath' and '$__logPath'.
 * $__rootLogPath Points the root directory that is dedicated for 7Zip's logging and reports.
 * $__logPath Points to the main place for all log files to be kept.

However, the already existing variable $__reportPath has been revised to now be the main stay for all of the report files instead of the root logging directory.