Created new function '__CreateDirectories()'
Added a new function named '__CreateDirectories()', which will automatically create the required directories necessary for this class to perfom the proper logging and reports.  This function will return $true if the directories either already exists or if the directories were created successfully, else - if there was a general failure during creating the directories then $false will be returned.