Added new function 'FileHash()'
Added a new function named 'FileHash()'.  This function will try to provide the hash value regarding a specific data-file.  This function, in addition, gives flexibility to use any of the supported hash algorithms provided in the Get-FileHash CMDLet.  Though, it is common to use MD5 and SHA1.
If an error occured or the file does not exist, null will be returned.