Resolved performance issues when Get-FileHash fails
Resolved an performance degrade issue when Get-FileHash returns an error but yet was still expected to get the hash property - despite it may sometimes never exist.  To resolve this issue, we merely try to cache the data to a new variable $hashInfo and then we would simply inspect the hash from that variable.  If we could not get the hash from Get-FileHash, then we would jump to the 'catch' condition immediately without needing to inspect a property.