Added a new function 'VerifyArchive()'
Added a new function named 'VerifyArchive()'.  This function will test the archive file to see if the archive data file is corrupted or damaged in anyway; useful to know if the archive is essentially usable or kaput.  In order to perform this test, this function utilizes 7Zip Integrity testing functionality with the 't' switch.