Minor touches && Added a new function 'ExtractArchive()'
Applied minor touches, nothing that affects the progra- just my OCD.

Added a new function named 'ExtractArchive()'.  This function provides the ability to extract any given supported archive file to the desired output path.  However, if the output directory already exists, this program will try to create a unique directory.  To make a unique directory when needed, and to make have more meaning, we will merely apply a timestamp to it.  This function will return true if the entire operation was a success, though if any error occurs - then false will be brought back instead.