Added a new function 'CreateArchive()'
Added a new function named 'CreateArchive()'.  This function will essentially create a new archive data file using a specific directory  to compact the contents while using the user's compression preferences to compact the files into the archive data file.  If incase, the archive file already exists with the same name, the archive file name will contain a time stamp to help make it both unique and provide some meaning to the name.  Though, if this is not enough to make the archive file unique - the archive file can not be created.

All of the user's settings are being appended as intended through a step-wise process.  This should make it easier to trace if something goes horribly wrong.