Added a new function 'CreateNewReport()'
Added a new function named 'CreateNewReport()'.  This function is designed to analyze the provided archive file and provide as much information as possible.  Such information includes: file properties (residing on the host's filesystem), file hash information, and a list of files that are are wrapped inside of the archive data file.  This functionality is adjacent to Git's Report, so it is possible to also get a PDF version upon request.