Removed variable '$__sourceBranch'
Removed a variable '$__sourceBranch'.  This variable was not yet used, but after much thought - it is not wise to have this variable or functionality.  The idea of this variable is to allow the program to automatically change branches for the user, however that alone can be bad idea as there could be files or changes that were not commited - thus losing work.  For ordinary users that only use the source to create builds - this is perfect, but for those developing within the sorce - it's a major issue.  This program will not perform such action.