Only generate a report if the user allows it.
I changed my mind on this variable, I thought it was not necessary - still kind-of unnecessary, but it's back for consistency sakes.  In order to generate a report, the variable '$__generateReport' must be true.  Otherwise, when the variable '$__generateReport' is false, the report will not be generated and the function 'CreateNewReport()' will return true.