Output builds and logs according to Project Name
Provide logs and generated builds in a directory that is dedicated to the project itself instead of unifying all different projects in one location.  For example, if working on Project 0 and Project 1 concurrently, do not mix them together when providing: logs, reports, and builds -- all projects should be neatly organized in their own filesystem, never thrown together.

This change should make it appear nicer when trying to keep projects separated and organized.  However, with this change, the program's boot-order has changed.  ProjectInformation must be the first to be instatiated before calling Initializations().