Added two new functions '__CreateDirectories()' && '__CheckRequiredDirectories()'
Added two new functions named the following: '__CheckRequiredDirectories()' and '__CreateDirectories()'.  The function, '__CheckRequiredDirectories()', will check to make sure that the required directories exist within the host's filesystem.  This function will return $true if all of the directories exist.
In addition, the function '__CreateDirectories()' will merely create the special directories.  This function will return $true, if the directories already exists or if the operation was successful.

Furthermore, revised the variable '$__rootLogPath' leaf name from 'dotNET_Compression' to 'PSArchive'.  Even though PowerShell uses the .NET Framework to use the Archive file operations, it is just better to say 'PowerShell Archive' or 'PSArchive'.