Added a new function ListFiles()
Added a new function named ListFiles().  This function will fetch all of the files that resides within the archive file and return the list depending what information is to be included.  It is possible to get a simple list that only contains a list of only the full file names, but it is also possible to get the technical information for each file that exists within the archive data file.