Fixed TargetDirectory var. validation in CreateArchive()
Fixed an issue where the validation would always fail if trying to check a target directory string that contains a wild-card.  For example: C:\Users\* or C:\Users\*.*.  To resolve this issue, merely filter out the wild-cards  using a new variable '$targetDirectoryFiltered' which contains a split-path version of the targetDirectory string (without the wild-card chars).