Resolve path issue that contains wildcards
Resolved an issue where the path string containing a wild-card would never work.  This was due to using the parameter -LiteralPath instead of -Path.  LiteralPath does not allow relative paths nor wildcards, but -Path will allow it.

This will be necessary to capture data from inside the leaf directory instead of leaf directory being the new root in the archive data file.