Added a new class LoadSaveUserConfiguration
Added a new class named 'LoadSaveUserConfiguration' (or Load\Save User Configurations).  This class's purpose (once implemented) will be to help guide save and also load a user configuration.  With this functionality, it will be possible to store all of the user's settings and preferences to a specific file, but to also load from the same specific file to restore the user's settings and preferences.  Thus, the user will not need to reconfigure the program after each new session, the last saved file will provide all of the details necessary for the user to pick back up where they left off.