[WIP] Objects in PowerShell are already passed by reference.
Despite it was never documented on Microsoft's documentation, objects are automatically passed by reference.  Because I was trying to pass the object by reference (and PS was doing the same), somehow things escalated in a bad way - which I could not easily resolve (if at all).  To resolve this, just pass the objects normally - PowerShell will automatically pass the objects by reference.

Removed some debugging code.

[WIP] Only the first object is able to get the proper data from the XML file.  Something is horribly wrong during the deserialization....