Added loading validation!
Implemented validation for each variable that is being loaded into the current session's environment.  This validation check is necessary to assure that each variable that is being loaded from the XML file into the program's current environment is 'correct'.  If the XML file was altered and contains faulty data, the program will use the current values while ignoring the misformed data that exists in the XML file.  Thus, this will help prevent any strange bugs\crashes from occurring out of the blue.