Added a new function AccessWebPage()
Added a new function named AccessWebpage().  This function will attempt to open the desired webpage using the user's preferred or default web browser.  If the URL Address is not a valid address or the page could not be loaded - this function will return false.  Otherwise, this function will return true if the operation was successful.

NOTE: In order to for the URL Address to be valid, it must contain either 'http://' or 'www.' as the first part of the string.