Directional Shift: GAME_ASSET_PATHING
Looking back at how TGRDM3 was compiled, it focused a lot on throwing everything into the main directories with no-organization.  With that, this compiler should do the same as well.  Meaning, that the output directories will be as 'parent', disallowing sub-directories.

To further elaborate, it will be possible to have a directory named: 'Textures' but not possible to have 'Textures/Foo/Bar'.

This shift I believe will have to reduce code complexity as well as to make the compiler easier to work with in the future for new features or maintainability.